+++ tests/integration/test_target_bigquery_stream.py
@@ -1,7 +1,6 @@
 import datetime
 import json
 import os
-import unittest
 import unittest.mock as mock
 from datetime import timezone
 from decimal import Decimal, getcontext
@@ -26,55 +25,10 @@ def query(bigquery, query):
     return [dict(row.items()) for row in result]
 
 
-class TestIntegration(unittest.TestCase):
+class TestIntegrationSchema(test_utils.TestIntegration):
     """
-    Integration Tests
+    Integration Tests about reading streams
     """
-    maxDiff = None
-
-    def setUp(self):
-        self.config = test_utils.get_test_config()
-        bigquery = DbSync(self.config)
-
-        # Drop target schema
-        if self.config['default_target_schema']:
-            bigquery.client.delete_dataset(
-                self.config['default_target_schema'],
-                delete_contents=True,
-                not_found_ok=True)
-
-    def persist_lines(self, lines):
-        """Loads singer messages into bigquery.
-        """
-        target_bigquery.persist_lines(self.config, lines)
-
-    def remove_metadata_columns_from_rows(self, rows):
-        """Removes metadata columns from a list of rows"""
-        d_rows = []
-        for r in rows:
-            # Copy the original row to a new dict to keep the original dict
-            # and remove metadata columns
-            d_row = r.copy()
-            for md_c in METADATA_COLUMNS:
-                d_row.pop(md_c, None)
-
-            # Add new row without metadata columns to the new list
-            d_rows.append(d_row)
-
-        return d_rows
-
-    def assert_metadata_columns_exist(self, rows):
-        """This is a helper assertion that checks if every row in a list has metadata columns"""
-        for r in rows:
-            for md_c in METADATA_COLUMNS:
-                self.assertTrue(md_c in r)
-
-    def assert_metadata_columns_not_exist(self, rows):
-        """This is a helper assertion that checks metadata columns don't exist in any row"""
-        for r in rows:
-            for md_c in METADATA_COLUMNS:
-                self.assertFalse(md_c in r)
-
     def assert_three_streams_are_into_bigquery(self, should_metadata_columns_exist=False,
                                                 should_hard_deleted_rows=False):
         """
