+++ tests/integration/utils.py
@@ -1,5 +1,6 @@
 import os
 import json
+import unittest
 
 
 def get_db_config():
@@ -40,3 +41,53 @@ def get_test_tap_lines(filename):
             lines.append(line)
 
     return lines
+
+class TestIntegration(unittest.TestCase):
+    """
+    Integration Tests
+    """
+    maxDiff = None
+
+    def setUp(self):
+        self.config = test_utils.get_test_config()
+        bigquery = DbSync(self.config)
+
+        # Drop target schema
+        if self.config['default_target_schema']:
+            bigquery.client.delete_dataset(
+                self.config['default_target_schema'],
+                delete_contents=True,
+                not_found_ok=True)
+
+    def persist_lines(self, lines):
+        """Loads singer messages into bigquery.
+        """
+        target_bigquery.persist_lines(self.config, lines)
+
+    def remove_metadata_columns_from_rows(self, rows):
+        """Removes metadata columns from a list of rows"""
+        d_rows = []
+        for r in rows:
+            # Copy the original row to a new dict to keep the original dict
+            # and remove metadata columns
+            d_row = r.copy()
+            for md_c in METADATA_COLUMNS:
+                d_row.pop(md_c, None)
+
+            # Add new row without metadata columns to the new list
+            d_rows.append(d_row)
+
+        return d_rows
+
+    def assert_metadata_columns_exist(self, rows):
+        """This is a helper assertion that checks if every row in a list has metadata columns"""
+        for r in rows:
+            for md_c in METADATA_COLUMNS:
+                self.assertTrue(md_c in r)
+
+    def assert_metadata_columns_not_exist(self, rows):
+        """This is a helper assertion that checks metadata columns don't exist in any row"""
+        for r in rows:
+            for md_c in METADATA_COLUMNS:
+                self.assertFalse(md_c in r)
+
